Alteracao dos ratings na api para inteiro.

+++ STrackerServer/STrackerServer/Controllers/Api/EpisodesRatingsController.cs
@@ -63,7 +63,7 @@ namespace STrackerServer.Controllers.Api
                 throw new HttpResponseException(HttpStatusCode.NotFound);
             }
 
-            return this.BaseGet(new { Rating = ratings.Average, Total = ratings.Ratings.Count });
+            return this.BaseGet(new { Rating = (int)ratings.Average, Total = ratings.Ratings.Count });
         }
 
         /// <summary>
